requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
flake8==3.8.4
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
isort==5.6.4
mccabe==0.6.1
multidict==4.7.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
zipp==3.4.0
